+++ spec/initializer_spec.rb
@@ -4,26 +4,25 @@ describe "extending all nodes in a folder through _init.rb" do
   end
 
   let!(:page) do
-    node "page.html.erb", parent: root,
-      source: %{I'm a page! This is a <%= node.show_test %>!}
+    root.create "page.html.erb",
+      source: %{I'm a page! This is a <%= page.root.test %>!}
   end
 
   let!(:initializer) do
     node "_init.rb", parent: root, source: <<-EOF
-extend_siblings do
-  on_setup do
-    @test = "test"
-  end
+on :created do
+  @test = "test"
+end
 
-  def show_test
-    @test
-  end
+def test
+  @test
 end
 EOF
   end
 
   specify "works :)" do
     root.stage!
+    expect(root.test).to eq("test")
     expect(page.render).to eq("I'm a page! This is a test!")
   end
 end