+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Version History
 
+### HEAD
+
+- **NEW:** Event system! (TODO: add link to tutorial.)
+
+
 ### 0.6.2 (2017-01-29)
 
 - **FIXED:** Images and other binaries would crash Flutterby when trying to extract frontmatter from them. Woops! ([#29](https://github.com/hmans/flutterby/issues/29))