+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Version History
 
-### HEAD
+### 0.7.0 (in development)
 
 - **NEW:** Event system! (TODO: add link to tutorial.)
 - **NEW:** Flutterby will now add your project's `lib` directory to Ruby's load path. Use it to house any Ruby modules that you want out of your site directory.